Revert "possible-fix: search not showing up in some devices #93"
This reverts commit e6460d84f523b801cfeaa431a03c2e6722e6e344.

+++ repository/src/main/java/app/simple/felicity/repository/models/Audio.java
@@ -190,7 +190,7 @@ public class Audio implements Parcelable {
         uri = in.readString();
         track = in.readInt();
         author = in.readString();
-        size = in.readLong();
+        size = in.readInt();
         albumArtist = in.readString();
         year = in.readString();
         bitrate = in.readLong();
